backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
datasets==4.0.0
dill==0.3.8
fastapi==0.116.1
filelock==3.19.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
Levenshtein==0.27.1
MarkupSafe==3.0.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.5
numpy==1.26.4
omegaconf==2.3.0
onnx==1.19.0
onnxruntime==1.22.1

packaging==25.0
pandas==2.3.2
pdf2image==1.17.0
phocr==1.0.3
pillow==11.3.0
propcache==0.3.2
protobuf==6.32.0
pyarrow==21.0.0
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.4
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.1
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scipy==1.13.1
sentencepiece==0.2.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
easyocr==1.7.2
imageio==2.37.0
lazy_loader==0.4
ninja==1.13.0
PyPDF2==3.0.1
python-bidi==0.6.6
scikit-image==0.25.2
tifffile==2025.9.20
nest_asyncio==1.6.0
